refactor: Changed intrinsic-function fixture name
