+++ test/fixtures/state-machines/intrinsic-functions/array/array.json
