+++ test/fixtures/state-machines/intrinsic-functions/array/empty-array.json
