+++ test/fixtures/state-machines/intrinsic-functions/array/strings.json
