+++ test/fixtures/state-machines/intrinsic-functions/format/bad-format-string.json
