+++ test/fixtures/state-machines/intrinsic-functions/format/everything.json
