+++ test/fixtures/state-machines/intrinsic-functions/format/format.json
