+++ test/fixtures/state-machines/intrinsic-functions/format/no-format-string.json
