+++ test/fixtures/state-machines/intrinsic-functions/index.js
