+++ test/fixtures/state-machines/intrinsic-functions/json-to-string/no-params.json
