+++ test/fixtures/state-machines/intrinsic-functions/json-to-string/number.json
