+++ test/fixtures/state-machines/intrinsic-functions/json-to-string/object.json
