+++ test/fixtures/state-machines/intrinsic-functions/json-to-string/param-is-not-a-path.json
