+++ test/fixtures/state-machines/intrinsic-functions/json-to-string/string.json
