+++ test/fixtures/state-machines/intrinsic-functions/string-to-json/no-params.json
