+++ test/fixtures/state-machines/intrinsic-functions/string-to-json/string-to-json.json
